+++ tests/e2e/fixtures/ActionRootMarginChangeFixture.svelte
@@ -9,7 +9,7 @@
   {intersecting ? "Element is in view" : "Element is not in view"}
   <button
     data-testid="shrink-root-margin"
-    on:click={() => (rootMargin = "-200px")}
+    onclick={() => (rootMargin = "-200px")}
   >
     Shrink root margin
   </button>
@@ -17,7 +17,7 @@
 
 <div
   use:intersect={{ rootMargin }}
-  on:observe={(e) => (intersecting = e.detail.isIntersecting)}
+  onobserve={(e) => (intersecting = e.detail.isIntersecting)}
 >
   Hello world
 </div>